backend/requirements.txt
asgiref==3.11.1
Django==6.0.2
django-cors-headers==4.9.0
djangorestframework==3.16.1
sqlparse==0.5.5

# Image handling
Pillow>=10.0.0

# PostgreSQL
psycopg2-binary>=2.9.9

# Environment variables
python-dotenv>=1.0.0

# Auth Token
djangorestframework>=3.14.0

# Optional: JWT for more secure authentication
PyJWT>=2.8.1

# PDF Export
reportlab>=4.0.0

# Excel Export
openpyxl>=3.0.0

